Display current user group owner on top of member list

+++ goobi-viewer-core/src/main/java/io/goobi/viewer/managedbeans/AdminBean.java
@@ -666,10 +666,12 @@ public class AdminBean implements Serializable {
                         break;
                     case "delete":
                         logger.trace("Deleting UserRole: {}", userRole);
-                        if (DataManager.getInstance().getDao().deleteUserRole(userRole)) {
-                            Messages.info("deletedSuccessfully");
-                        } else {
-                            Messages.error("deleteFailure");
+                        if(userRole.getId() != null) {                            
+                            if (DataManager.getInstance().getDao().deleteUserRole(userRole)) {
+                                Messages.info("deletedSuccessfully");
+                            } else {
+                                Messages.error("deleteFailure");
+                            }
                         }
                         break;
                     default:
